Making deletion an env var

+++ src/adapters/slp-indexer/lib/db-backup.js
@@ -101,9 +101,12 @@ class DbBackup {
         `zip -r ${dbDir}/zips/slp-indexer-${height}.zip ${dbDir}/current`
       )
 
-      // Delete the old backup.
-      const oldHeight = height + epoch
-      this.shell.rm(`${dbDir}/zips/slp-indexer-${oldHeight}.zip`)
+      const deleteBackup = parseInt(process.env.DELETE_BACKUP)
+      if (deleteBackup) {
+        // Delete the old backup.
+        const oldHeight = height + epoch
+        this.shell.rm(`${dbDir}/zips/slp-indexer-${oldHeight}.zip`)
+      }
 
       // Reopen the databases.
       console.log('Reopening database')
